Refactor Engine components to be optional
